# internal/note/handlers.go
package note

import (
	"encoding/base64"
	"encoding/json"
	"errors"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"

	"github.com/jorgebuitragor/logday-server/internal/auth"
	"github.com/jorgebuitragor/logday-server/internal/db"
	"github.com/jorgebuitragor/logday-server/internal/realtime"
)

// Handler exposes the note HTTP endpoints, protected by authHandler's
// RequireAuth middleware.
type Handler struct {
	store    *store
	auth     *auth.Handler
	realtime *realtime.Hub
}

// NewHandler builds a Handler backed by s, notifying hub of every
// successful write.
func NewHandler(s *store, authHandler *auth.Handler, hub *realtime.Hub) *Handler {
	return &Handler{store: s, auth: authHandler, realtime: hub}
}

// Routes registers the note-related endpoints on r. Content lives on
// its own endpoint (POST .../content), separate from the LWW-governed
// fields below — see specs/arquitectura-inicial, "Resolución de
// conflictos".
func (h *Handler) Routes(r chi.Router) {
	r.Group(func(r chi.Router) {
		r.Use(h.auth.RequireAuth)

		r.Post("/notes", h.create)
		r.Patch("/notes/{id}", h.patch)
		r.Post("/notes/{id}/content", h.postContent)
		r.Delete("/notes/{id}", h.delete)
		r.Get("/notes", h.list)
	})
}

type noteRequest struct {
	ID        string    `json:"id"`
	Title     string    `json:"title"`
	Folder    string    `json:"folder"`
	Tags      []string  `json:"tags"`
	Created   string    `json:"created"`
	Updated   string    `json:"updated"`
	Pinned    bool      `json:"pinned"`
	UpdatedAt time.Time `json:"updated_at"`
}

func (req noteRequest) toNote(userID string) Note {
	tags := req.Tags
	if tags == nil {
		tags = []string{}
	}
	return Note{
		ID: req.ID, UserID: userID, Title: req.Title, Folder: req.Folder, Tags: tags,
		Created: req.Created, Updated: req.Updated, Pinned: req.Pinned, UpdatedAt: req.UpdatedAt,
	}
}

func validateNoteRequest(req noteRequest) error {
	if req.ID == "" {
		return errors.New("id is required")
	}
	if req.Title == "" {
		return errors.New("title is required")
	}
	if req.Created == "" {
		return errors.New("created is required")
	}
	if req.UpdatedAt.IsZero() {
		return errors.New("updated_at is required")
	}
	return nil
}

func (h *Handler) create(w http.ResponseWriter, r *http.Request) {
	userID, _ := auth.UserIDFromContext(r.Context())

	var req noteRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "invalid request body", http.StatusBadRequest)
		return
	}
	if err := validateNoteRequest(req); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	n := req.toNote(userID)
	h.upsert(w, r, &n)
}

// parseNotePatch reads a PATCH body's raw fields into a Patch.
func parseNotePatch(raw map[string]json.RawMessage) (Patch, error) {
	var patch Patch
	var err error
	if patch.Title, err = db.PatchField[string](raw, "title"); err != nil {
		return Patch{}, err
	}
	if patch.Folder, err = db.PatchField[string](raw, "folder"); err != nil {
		return Patch{}, err
	}
	if patch.Tags, err = db.PatchField[[]string](raw, "tags"); err != nil {
		return Patch{}, err
	}
	if patch.Created, err = db.PatchField[string](raw, "created"); err != nil {
		return Patch{}, err
	}
	if patch.Updated, err = db.PatchField[string](raw, "updated"); err != nil {
		return Patch{}, err
	}
	if patch.Pinned, err = db.PatchField[bool](raw, "pinned"); err != nil {
		return Patch{}, err
	}
	return patch, nil
}

func validateNotePatch(patch Patch) error {
	if patch.Title.Set && patch.Title.Value == "" {
		return errors.New("title cannot be empty")
	}
	if patch.Created.Set && patch.Created.Value == "" {
		return errors.New("created cannot be empty")
	}
	return nil
}

func (h *Handler) patch(w http.ResponseWriter, r *http.Request) {
	userID, _ := auth.UserIDFromContext(r.Context())
	id := chi.URLParam(r, "id")

	raw, err := db.ParsePatch(r.Body)
	if err != nil {
		http.Error(w, "invalid request body", http.StatusBadRequest)
		return
	}

	updatedAt, err := db.PatchField[time.Time](raw, "updated_at")
	if err != nil || !updatedAt.Set || updatedAt.Value.IsZero() {
		http.Error(w, "updated_at is required", http.StatusBadRequest)
		return
	}

	patch, err := parseNotePatch(raw)
	if err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}
	if err := validateNotePatch(patch); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	stored, changed, err := h.store.patchNote(r.Context(), id, userID, patch, updatedAt.Value)
	if err != nil {
		switch {
		case errors.Is(err, errNotFound):
			http.Error(w, "note not found", http.StatusNotFound)
		case errors.Is(err, errForbidden):
			http.Error(w, "note belongs to another user", http.StatusForbidden)
		default:
			http.Error(w, "internal error", http.StatusInternalServerError)
		}
		return
	}
	if changed {
		h.realtime.Notify(stored.UserID, "note", stored.ID, stored.Seq)
	}
	writeJSON(w, http.StatusOK, stored)
}

func (h *Handler) upsert(w http.ResponseWriter, r *http.Request, n *Note) {
	stored, err := h.store.upsertNote(r.Context(), n)
	if err != nil {
		switch {
		case errors.Is(err, errForbidden):
			http.Error(w, "note belongs to another user", http.StatusForbidden)
		case errors.Is(err, errConflict):
			http.Error(w, "a newer version of this note already exists", http.StatusConflict)
		default:
			http.Error(w, "internal error", http.StatusInternalServerError)
		}
		return
	}
	h.realtime.Notify(stored.UserID, "note", stored.ID, stored.Seq)
	writeJSON(w, http.StatusOK, stored)
}

type contentRequest struct {
	ContentUpdate string    `json:"content_update"`
	UpdatedAt     time.Time `json:"updated_at"`
}

func (h *Handler) postContent(w http.ResponseWriter, r *http.Request) {
	userID, _ := auth.UserIDFromContext(r.Context())
	id := chi.URLParam(r, "id")

	var req contentRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "invalid request body", http.StatusBadRequest)
		return
	}
	if req.ContentUpdate == "" {
		http.Error(w, "content_update is required", http.StatusBadRequest)
		return
	}
	if req.UpdatedAt.IsZero() {
		http.Error(w, "updated_at is required", http.StatusBadRequest)
		return
	}

	update, err := base64.StdEncoding.DecodeString(req.ContentUpdate)
	if err != nil {
		http.Error(w, "content_update must be base64-encoded", http.StatusBadRequest)
		return
	}

	stored, err := h.store.applyContentUpdate(r.Context(), id, userID, update, req.UpdatedAt)
	if err != nil {
		switch {
		case errors.Is(err, errNotFound):
			http.Error(w, "note not found", http.StatusNotFound)
		case errors.Is(err, errForbidden):
			http.Error(w, "note belongs to another user", http.StatusForbidden)
		default:
			http.Error(w, "internal error", http.StatusInternalServerError)
		}
		return
	}
	h.realtime.Notify(stored.UserID, "note", stored.ID, stored.Seq)
	writeJSON(w, http.StatusOK, stored)
}

func (h *Handler) delete(w http.ResponseWriter, r *http.Request) {
	userID, _ := auth.UserIDFromContext(r.Context())
	id := chi.URLParam(r, "id")

	seq, err := h.store.softDelete(r.Context(), id, userID)
	if err != nil {
		switch {
		case errors.Is(err, errNotFound):
			http.Error(w, "note not found", http.StatusNotFound)
		case errors.Is(err, errForbidden):
			http.Error(w, "note belongs to another user", http.StatusForbidden)
		default:
			http.Error(w, "internal error", http.StatusInternalServerError)
		}
		return
	}
	h.realtime.Notify(userID, "note", id, seq)
	w.WriteHeader(http.StatusNoContent)
}

func (h *Handler) list(w http.ResponseWriter, r *http.Request) {
	userID, _ := auth.UserIDFromContext(r.Context())

	notes, err := h.store.listNotes(r.Context(), userID)
	if err != nil {
		http.Error(w, "internal error", http.StatusInternalServerError)
		return
	}
	if notes == nil {
		notes = []Note{}
	}
	writeJSON(w, http.StatusOK, notes)
}