# openapi.yaml
openapi: 3.0.3
info:
  title: Logday Server API
  description: |
    API de sync self-hosted para [Logday](https://github.com/jorgebuitragor/logday),
    al estilo Vaultwarden/Bitwarden. Un cliente (desktop, web, móvil, extensión)
    apunta a la URL de una instancia self-hosted y usa esta API para sincronizar
    sus datos entre dispositivos.

    ## Modelo de autenticación

    - No hay auto-registro. Los usuarios los crea un admin (`POST /admin/users`
      o el panel web), salvo el primer admin de la instancia (`/setup` o
      variables de entorno `ADMIN_EMAIL`/`ADMIN_PASSWORD`, ver README).
    - `POST /auth/login` devuelve un `access_token` (JWT de corta duración,
      usado como `Authorization: Bearer <token>` en cada request) y un
      `refresh_token` (opaco, de larga duración, ligado a un `device_id`).
    - Cuando el access token expira, `POST /auth/refresh` cambia el refresh
      token (rotación: el anterior queda invalidado; reusarlo revoca el
      dispositivo — indicio de robo de token).
    - Los TTLs de ambos tokens y el límite de dispositivos por usuario son
      configurables por el admin desde el panel (`internal/settings`), por lo
      que no son constantes fijas — un cliente no debe asumir valores.

    ## Modelo de sync

    Cada recurso (`tasks`, `notes`, `overtime-entries`, `overtime-month-meta`,
    `calendar-events`, `absence-days`, `daily-entries`) expone su propio CRUD
...
    típico:

...
       cursor local.
...

...
    endpoint de contenido en base64.

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /notes:
    get:
      operationId: listNotes
      tags: [Notes]
      summary: Lista todas las notas del usuario
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: upsertNoteByBody
      tags: [Notes]
      summary: Crea o actualiza los metadatos de una nota (upsert por `id` del body)
      description: |
        Este endpoint sólo escribe los campos LWW (título, folder, tags,
        pinned, etc.) — **nunca** toca el contenido. El contenido se escribe
        exclusivamente vía `POST /notes/{id}/content`, que es CRDT y no
        participa del chequeo de `updated_at`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteWrite"
      responses:
        "200":
          description: Nota creada o actualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: La nota pertenece a otro usuario.
          content:
            text/plain: { schema: { type: string } }
        "409":
          description: Conflicto LWW en los metadatos.
          content:
            text/plain: { schema: { type: string } }
  /notes/{id}:
    patch:
      operationId: patchNote
      tags: [Notes]
      summary: Edita campos puntuales de los metadatos de una nota (LWW por campo)
      description: |
        Mismo protocolo que `PATCH /tasks/{id}` (ver `specs/lww-por-campo`):
        solo los campos presentes se evalúan, cada uno se resuelve por su
        propio timestamp, nunca responde `409`. Nunca toca el contenido — eso
        sigue siendo exclusivo de `POST /notes/{id}/content`.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotePatch"
      responses:
        "200":
          description: Estado actual de la nota tras aplicar los campos que ganaron el LWW.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: La nota pertenece a otro usuario.
          content:
            text/plain: { schema: { type: string } }
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteNote
      tags: [Notes]
      summary: Soft-delete de una nota (metadatos y contenido)
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Borrada (tombstone).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: La nota pertenece a otro usuario.
          content:
            text/plain: { schema: { type: string } }
        "404":
          $ref: "#/components/responses/NotFound"
  /notes/{id}/content:
    post:
      operationId: updateNoteContent
      tags: [Notes]
      summary: Mergea un update CRDT del contenido de la nota
      description: |
        A diferencia de `PATCH /notes/{id}`, esto **nunca rechaza por
        antigüedad**: los updates CRDT conmutan y son idempotentes, así que
        siempre se aplican y mergean con el estado guardado, sin importar el
        orden de llegada entre dispositivos. `content_update` es un update
        binario (formato Yjs) codificado en base64 — no el texto completo.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdateRequest"
      responses:
        "200":
          description: Update aplicado — devuelve la nota completa con el contenido ya mergeado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Body inválido, falta `content_update`/`updated_at`, o `content_update` no es base64 válido.
          content:
            text/plain: { schema: { type: string } }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: La nota pertenece a otro usuario.
          content:
            text/plain: { schema: { type: string } }
        "404":
          $ref: "#/components/responses/NotFound"
...
  /absence-days:
    get:
      operationId: listAbsenceDays
      tags: [Absence]
      summary: Lista todos los días de ausencia del usuario
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AbsenceDay"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: upsertAbsenceDayByBody
      tags: [Absence]
      summary: Crea o actualiza un día de ausencia (upsert por `id` del body)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbsenceDayWrite"
      responses:
        "200":
          description: Día creado o actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbsenceDay"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: El día pertenece a otro usuario.
          content:
            text/plain: { schema: { type: string } }
        "409":
          description: Conflicto LWW.
          content:
            text/plain: { schema: { type: string } }
  /absence-days/{id}:
    patch:
      operationId: patchAbsenceDay
      tags: [Absence]
      summary: Edita campos puntuales de un día de ausencia (LWW por campo)
      description: Mismo protocolo que `PATCH /tasks/{id}` (ver `specs/lww-por-campo`) — solo los campos presentes se evalúan, cada uno por su propio timestamp, nunca responde `409`.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbsenceDayPatch"
      responses:
        "200":
          description: Estado actual del día tras aplicar los campos que ganaron el LWW.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbsenceDay"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: El día pertenece a otro usuario.
          content:
            text/plain: { schema: { type: string } }
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteAbsenceDay
      tags: [Absence]
      summary: Soft-delete de un día de ausencia
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Borrado (tombstone).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: El día pertenece a otro usuario.
          content:
            text/plain: { schema: { type: string } }
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token devuelto por `/auth/login` o `/auth/refresh`, enviado como `Authorization: Bearer <access_token>`."
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      description: ID generado por el cliente (UUID recomendado) al crear el recurso.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Body inválido o falla una validación de campo (ver el mensaje).
      content:
        text/plain:
          schema:
            type: string
    Unauthorized:
      description: Falta el header `Authorization`, o el access token es inválido/expiró.
      content:
        text/plain:
          schema:
            type: string
            example: invalid or expired token
...
    NotFound:
      description: El recurso no existe (o ya estaba borrado) para este usuario.
      content:
        text/plain:
          schema:
            type: string
  schemas:
...
    NoteWrite:
      type: object
      required: [id, title, created, updated_at]
      properties:
        id:
          type: string
        title:
          type: string
        folder:
          type: string
        tags:
          type: array
          items:
            type: string
        created:
          type: string
        updated:
          type: string
        pinned:
          type: boolean
        updated_at:
          type: string
          format: date-time
    Note:
      allOf:
        - $ref: "#/components/schemas/NoteWrite"
        - type: object
          properties:
            content:
              type: string
              description: Texto plano derivado del estado CRDT — de solo lectura, se escribe vía `POST /notes/{id}/content`.
            content_state:
              type: string
              description: Estado CRDT completo (formato Yjs) codificado en base64 — de solo lectura, útil para que el cliente lo cargue directo en su propio doc CRDT en vez de reconstruirlo de `content`.
            seq:
              type: integer
              format: int64
            deleted_at:
              type: string
              format: date-time
              nullable: true
    NotePatch:
      type: object
      required: [updated_at]
      description: Todos los campos son opcionales salvo `updated_at` (ver `PATCH /notes/{id}`). No incluye contenido — eso es `ContentUpdateRequest`.
      properties:
        title:
          type: string
        folder:
          type: string
        tags:
          type: array
          items:
            type: string
        created:
          type: string
        updated:
          type: string
        pinned:
          type: boolean
        updated_at:
          type: string
          format: date-time
    ContentUpdateRequest:
      type: object
      required: [content_update, updated_at]
      properties:
        content_update:
          type: string
          format: byte
          description: Update CRDT binario (formato Yjs) codificado en base64. No es el texto completo — es un delta que se mergea con el estado guardado.
        updated_at:
          type: string
          format: date-time

    # --- Overtime ---
...
    AbsenceDayWrite:
      type: object
      required: [id, date, type, updated_at]
      properties:
        id:
          type: string
        date:
          type: string
          format: date
        type:
          type: string
          enum: [incapacidad, vacaciones, otro]
        note:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
    AbsenceDay:
      allOf:
        - $ref: "#/components/schemas/AbsenceDayWrite"
        - type: object
          properties:
            seq:
              type: integer
              format: int64
            deleted_at:
              type: string
              format: date-time
              nullable: true
    AbsenceDayPatch:
      type: object
      required: [updated_at]
      description: Todos los campos son opcionales salvo `updated_at` (ver `PATCH /absence-days/{id}`). `note` enviado como `null` explícito lo limpia.
      properties:
        date:
          type: string
          format: date
        type:
          type: string
          enum: [incapacidad, vacaciones, otro]
        note:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time

    # --- Daily entries ---
...